Moved formatDate to a helper function file within script folder
